ajustando servicos de cadstro e painel de titulos para se comportar com o usuario setado

+++ src/main/java/com/gestor/cobranca/service/CadastroTituloService.java
@@ -19,8 +19,9 @@ public class CadastroTituloService {
 	private Titulos titulos;
 	
 	
-	public void salvar(Titulo titulo) throws IllegalAccessException {
+	public void salvar(Titulo titulo,Usuario usuario) throws IllegalAccessException {
 		try {
+			titulo.setUsuario(usuario);
 			titulos.save(titulo);	
 		}catch (DataIntegrityViolationException e) {
 			throw new IllegalAccessException("Formato de data inválido");
@@ -28,21 +29,29 @@ public class CadastroTituloService {
 		
 	}
 	
-	public Titulo editar(Long codigoTitulo) {
+	public Titulo editar(Long codigoTitulo,Usuario usuario) {
 		Titulo titulo = titulos.getOne(codigoTitulo);
+		titulo.setUsuario(usuario);
 		return titulo;
 	}
 	
-	public void excluir(Long codigo) {
-		titulos.deleteById(codigo);
+	public void excluir(Long codigo,Usuario usuario) {
+		Titulo titulo = titulos.getOne(codigo);
+		if (titulo.getUsuario().getId().equals(usuario.getId())){
+			titulos.deleteById(codigo);
+		}
 	}
 
-	public String receber(Long codigo) {
+	public String receber(Long codigo, Usuario usuario) {
 		Titulo titulo = titulos.getOne(codigo);
-		titulo.setStatus(StatusTitulo.RECEBIDO);
-		titulos.save(titulo);
-		
-		return StatusTitulo.RECEBIDO.getDescricao();
+
+		if (titulo.getUsuario().getId().equals(usuario.getId())){
+			titulo.setStatus(StatusTitulo.RECEBIDO);
+			titulos.save(titulo);
+
+			return StatusTitulo.RECEBIDO.getDescricao();
+		}
+		return StatusTitulo.PENDENTE.getDescricao();
 	}
 	
 	public List<Titulo> filtrar(TituloFilter filtro, Usuario usuario){
